0-linear.c

+++ 0x1E-search_algorithms/0-linear.c
@@ -10,14 +10,14 @@
 
 int linear_search(int *array, size_t size, int value)
 {
-	int i = 0;
+	size_t i = 0;
 
 	/* Check if the input array is NULL */
 	if (array == NULL)
 		return (-1);
 
 	/* iterate through the array */
-	for (i = 0; i < (int)size; i++)
+	for (i = 0; i < size; i++)
 	{
 		/* print the comparison being made */
 		printf("Value checked array[%lu] = [%d]\n", i, array[i]);
